+++ packages/typespec-rust/src/codemodel/customizations.ts
@@ -4,7 +4,7 @@
 *--------------------------------------------------------------------------------------------*/
 
 /** ModelFieldCustomizations defines the set of customizations for model fields. */
-export type ModelFieldCustomizations = DeserializeWith;
+export type ModelFieldCustomizations = DeserializeWith | SerializeWith;
 
 /** DeserializeWith indicates a field should use the specified deserializer function. */
 export interface DeserializeWith {
@@ -14,12 +14,27 @@ export interface DeserializeWith {
   name: string;
 }
 
+/** SerializeWith indicates a field should use the specified serializer function. */
+export interface SerializeWith {
+  kind: 'serializeWith';
+
+  /** the fully qualified name of the serializer function */
+  name: string;
+}
+
 ///////////////////////////////////////////////////////////////////////////////////////////////////
 ///////////////////////////////////////////////////////////////////////////////////////////////////
 
 export class DeserializeWith {
   constructor(name: string) {
     this.kind = 'deserializeWith';
+    this.name = name;
+  }
+}
+
+export class SerializeWith {
+  constructor(name: string) {
+    this.kind = 'serializeWith';
     this.name = name;
   }
 }